refactor:FSD 구조 개편 (searchbar -> search)
